src/requirements.txt
# Azure Functions for Python
azure-functions

# Azure Key Vault + Managed Identity
azure-identity
azure-keyvault-secrets

# HTTP requests library
requests>=2.31.0

# Additional dependencies
certifi
pydantic>=2.0.0
